Auth fixes v3

+++ app/Http/Controllers/Auth/AuthController.php
@@ -56,10 +56,17 @@ class AuthController extends Controller
 
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
+        if ($request->user()){
+            $request->user()->currentAccessToken()->delete();
+            return response()->json([
+                'message' => 'Logged out.'
+            ], 200);
+        }
 
         return response()->json([
-            'message' => 'Logged out.'
-        ], 200);
+        'message' => 'Login first.'
+    ], 403);
+
+
     }
 }
